requirements.txt
matplotlib
scipy
torch==0.4.1
torchvision==0.2.1
tqdm==4.11.2
tensorboardX
pyyaml
pydensecrf
protobuf
jinja2
IPython
